crud title, reimbursement

+++ API/Models/Title.cs
@@ -8,7 +8,7 @@ namespace API.Models
         [Key]
         public string Id_Title { get; set; }
         public string Title_Name { get; set; }
-        public int Reimburse_Limit { get; set; }
+        public float Reimburse_Limit { get; set; }
 
         public ICollection<AccountDetail> AccountDetails { get; set; }
     }
